+++ app/src/main/java/pl/rethagos/musicplayer/recyclerview/OnAudioFileClickListener.kt
@@ -1,0 +1,7 @@
+package pl.rethagos.musicplayer.recyclerview
+
+import pl.rethagos.musicplayer.model.AudioFile
+
+interface OnAudioFileClickListener {
+    fun onClick(audioFile: AudioFile?)
+}